testing for bugs

+++ src/com/ftpserver/logger/NetTransferLogger.java
@@ -8,8 +8,8 @@ import com.ftpserver.Statics;
 public class NetTransferLogger {
 
     private static NetTransferLogger _instance;
-    private double upsum = 0;
-    private double downsum = 0;
+    private long upsum = 0;
+    private long downsum = 0;
 
     private NetTransferLogger() {
     }
@@ -24,8 +24,20 @@ public class NetTransferLogger {
     public void logNetTransfer(Integer length, Statics.NET_TRANSFER_TYPE type) {
         if (type == Statics.NET_TRANSFER_TYPE.DOWNLOAD) {
             downsum += length;
+//            System.out.println();
+//            System.out.println("============================================");
+//            System.out.println("= LOG: " + length + " BYTES SENT THIS TIME");
+//            System.out.println("= LOG: " + downsum + " BYTES IN TOTAL");
+//            System.out.println("============================================");
+//            System.out.println();
         } else if (type == Statics.NET_TRANSFER_TYPE.UPLOAD) {
             upsum += length;
+//            System.out.println();
+//            System.out.println("============================================");
+//            System.out.println("= LOG: " + length + " BYTES RECVED THIS TIME");
+//            System.out.println("= LOG: " + upsum + " BYTES IN TOTAL");
+//            System.out.println("============================================");
+//            System.out.println();
         }
     }
